Initialize rename of repository to solarspatialtools. Github and pip renaming pending.
