+++ Changelog.md
@@ -59,3 +59,4 @@ site itself.
 - Demo for digitizing a plant equipment layout added by @williamhobbs
 - Improve comments in `cmv_demo`
 - Add sphinx documentation for entire library
+- Name change to SolarSpatialTools
