+++ src/solarspatialtools/irradiance.py
