+++ src/solarspatialtools/spatial.py
