+++ src/solarspatialtools/stats.py
